# swagger.yaml
openapi: 3.0.0
info:
  title: Chatbot Backend Service
  description: API documentation for the Chatbot Backend Service
  version: 1.0.0

servers:
  - url: http://localhost:5000
    description: Local server

paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: testuser
                password:
                  type: string
                  example: testpassword
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created successfully
        '400':
          description: Bad request

  /login:
    post:
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: testuser
                password:
                  type: string
                  example: testpassword
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: your_jwt_token
        '401':
          description: Login failed

  /chat:
    post:
      summary: Send a message to the chatbot
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  example: Hello
      responses:
        '200':
          description: Chatbot response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    example: This is a mock response

  /history:
    get:
      summary: Get conversation history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Conversation history
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        message:
                          type: string
                          example: Hello
                        response:
                          type: string
                          example: This is a mock response
                        timestamp:
                          type: string
                          format: date-time
                          example: 2024-07-05T12:34:56

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT